Juno updates including - Copyright date to 2022 - StandardMem interface instead of SimpleMem - New naming conventions - Fix compiler warnings

+++ asm/asmop.h
@@ -1,8 +1,8 @@
-// Copyright 2013-2021 NTESS. Under the terms
+// Copyright 2013-2022 NTESS. Under the terms
 // of Contract DE-NA0003525 with NTESS, the U.S.
 // Government retains certain rights in this software.
 //
-// Copyright (c) 2013-2021, NTESS
+// Copyright (c) 2013-2022, NTESS
 // All rights reserved.
 //
 // Portions are copyright of other developers:
@@ -83,7 +83,7 @@ public:
 		}
 
 		char regNum[JUNO_MAX_REGISTER_DIGITS];
-		for(int i = 1; i < std::max(static_cast<size_t>(JUNO_MAX_REGISTER_DIGITS), strlen(regst)); ++i) {
+		for(size_t i = 1; i < std::max(static_cast<size_t>(JUNO_MAX_REGISTER_DIGITS), strlen(regst)); ++i) {
 			regNum[i-1] = regst[i];
 		}
